backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
feedparser==6.0.11
python-dateutil==2.8.2
apscheduler==3.10.4
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
pydantic==2.6.3